// src/app/apps/categoria2-container/models/categoria2.ts
export interface Categorias2 {
  categoria2_id: number;
  categoria2_codigo: string;
  categoria2_descripcion: string;
  categoria2_descripcion_corta: string;
  categoria2_cat: number;
  categoria2: Categoria2[];
}

export class Categoria2 {
  categoria2_id: number;
  categoria2_codigo: string;
  categoria2_descripcion: string;
  categoria2_descripcion_corta: string;
  categoria2_cat: number;
}